Add searching

+++ src/main/java/sample/service/PersonService.java
@@ -19,24 +19,31 @@ public class PersonService {
 	public PersonService(PersonRepository personRepository) {
 		this.personRepository = personRepository;
 	}
-	
+
 	public List<Person> findAll() {
 		List<Person> people = new ArrayList<>();
-		for(Person person: personRepository.findAll()){
+		for (Person person : personRepository.findAll()) {
 			people.add(person);
 		}
 		return people;
-}
+	}
 
 	public void save(Person person) {
 		personRepository.save(person);
 	}
-	
+
 	public Person findPerson(int id) {
 		return personRepository.findOne(id);
 	}
-	
+
 	public void delete(int id) {
-		personRepository.delete(id);
+		if (personRepository.exists(id)) {
+			personRepository.delete(id);
+		}
+	}
+
+	public List<Person> search(String str) {
+		//return personRepository.findById(Integer.parseInt(str));
+		return personRepository.findByFirstNameContainingOrLastNameContainingOrBirthDateContainingOrEmailContainingOrPhoneContaining(str, str, str, str, str);
 	}
 }
